Refactor and fix jira_get_issue bugs #294, #295 (#299)
* refactor: init protocols and remove duplicate

- Remove duplicate `get_issue` in `EpicsMixin` (found in `IssuesMixin`)
- Add protocols to give better static checks

* fix: prevent returning comments when default

Prevent `jira_get_issue` from returning comments in fields
when fields does not contain "comment".

Refs: #295

* feat: add JiraChangelog

- New `models/jira/changelog.py`
- Add Changelog handlings to
  - `JiraIssue.from_api_response`
  - `JiraIssue.to_simplified_dict`

Refs: #294

* test: fix test_issues.py calling

It is supposed to not include comments as default
in tool's argument description.

* test: fix unwanted comment fetching in tests

Fix tests from asserting fetching comments when "comment" not in fields

fix: fix get_issue not fetch comments

Fix get_issue() not calling _get_issue_comments_if_needed when
comment_limit is None.

* fix: fix get_project_issues return type hint

* fix: remove duplicate get_project_issues

It is duplicated with the same function in ProjectMixins.

* test: move and remove some tests

In `test_projects.py`:
- Add `projects_mixin_with_search` as a blended mixin
- Add `test_get_project_issues` and `test_get_project_issues`
  from `test_search.py`
- Remove `test_get_project_issues_without_search_mixin` as unnecessary
- Fix `test_get_project_issues_invalid_response` and
  `test_get_project_issues_exception`

In `test_search.py`:
- Remove `test_get_project_issues` and `test_get_project_issues`
  (move to `test_projects.py`)
- Fix `test_get_board_issues` function documentation

* test: remove test_get_epic_issues_no_search_issues

Remove as unnecessary.

* test: add ProjectsMixinWithSearch

Protocols are need to be implemented before instantiation.